fast_api_demo_project/requirements.txt
aiomysql==0.2.0
aiosqlite==0.21.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.7
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
MouseInfo==0.1.3
mysql-connector-python==9.2.0
outcome==1.3.0.post0
packaging==24.2
passlib==1.7.4
pillow==11.1.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.1
requests==2.32.3
selenium==4.28.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.37
starlette==0.45.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.34.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
